# Accessing isArray method is relatively expensive operation