--- a/source/_internals/_isArray.js
+++ b/source/_internals/_isArray.js
@@ -1,0 +1,2 @@
+// extract for optimization
+export const _isArray = Array.isArray