--- a/source/filterAsync.js
+++ b/source/filterAsync.js
@@ -1,3 +1,4 @@
+import { _isArray } from './_internals/_isArray'
 import { filter } from './filter'
 import { mapAsync } from './mapAsync'
 
@@ -9,7 +10,7 @@ export function filterAsync(predicate, listOrObject){
   return new Promise((resolve, reject) => {
     mapAsync(predicate, listOrObject)
       .then(predicateResult => {
-        if (Array.isArray(predicateResult)){
+        if (_isArray(predicateResult)){
           const filtered = listOrObject.filter((_, i) => predicateResult[ i ])
 
           return resolve(filtered)