--- a/source/flatten.js
+++ b/source/flatten.js
@@ -1,8 +1,10 @@
+import { _isArray } from './_internals/_isArray'
+
 export function flatten(list, input){
   const willReturn = input === undefined ? [] : input
 
   for (let i = 0; i < list.length; i++){
-    if (Array.isArray(list[ i ])){
+    if (_isArray(list[ i ])){
       flatten(list[ i ], willReturn)
     } else {
       willReturn.push(list[ i ])