--- a/source/mapAsync.js
+++ b/source/mapAsync.js
@@ -1,5 +1,7 @@
+import { _isArray } from './_internals/_isArray'
+
 async function mapAsyncFn(fn, arr){
-  if (Array.isArray(arr)){
+  if (_isArray(arr)){
     const willReturn = []
     let i = 0
     for (const a of arr){