Thoughts about stages you can read here #475 
Not all perf benchs showed. Processing power in some other function will speed up (transpose for ex.).

## Initalization stage
### filter (empty array):
**Before**
```js
  Rambda:
    5 045 006 ops/s, ±0.96%     | 95.58% slower
  Ramda:
    1 697 448 ops/s, ±1.48%     | slowest, 98.51% slower
  Lodash:
    114 056 063 ops/s, ±2.48%   | fastest
```
**After**
```js
  Rambda:
    287 421 997 ops/s, ±1.15%   | fastest
  Ramda:
    1 727 667 ops/s, ±1.27%     | slowest, 99.4% slower
  Lodash:
    120 056 085 ops/s, ±2.81%   | 58.23% slower
```

### clone (empty object):
**Before**
```js
  Rambda:
    5 006 674 ops/s, ±0.72%   | fastest
  Ramda:
    1 310 727 ops/s, ±2.16%   | slowest, 73.82% slower
  Lodash.cloneDeep:
    1 371 105 ops/s, ±1.97%   | 72.61% slower
```
**After**
```js
  Rambda:
    144 949 647 ops/s, ±1.19%   | fastest
  Ramda:
    1 264 141 ops/s, ±1.66%     | 99.13% slower
  Lodash.cloneDeep:
    1 090 004 ops/s, ±1.80%     | slowest, 99.25% slower
```
### map (empty array):
**Before**
```js
  Rambda:
    2 630 058 ops/s, ±0.81%   | 45.76% slower
  Ramda:
    1 244 835 ops/s, ±1.24%   | slowest, 74.33% slower
  Lodash:
    4 848 991 ops/s, ±1.22%   | fastest
```
**After**
```js
  Rambda:
    4 869 840 ops/s, ±1.14%   | fastest
  Ramda:
    1 207 391 ops/s, ±1.27%   | slowest, 75.21% slower
  Lodash:
    4 569 654 ops/s, ±1.23%   | 6.16% slower
```
### includes (empty array):
**Before**
```js
  Rambda:
    1 627 034 ops/s, ±1.27%   | slowest, 56.43% slower
  Ramda:
    3 734 149 ops/s, ±2.38%   | fastest
```
**After**
```js
  Rambda:
    34 637 306 ops/s, ±3.03%   | fastest
  Ramda:
    3 587 805 ops/s, ±2.41%    | slowest, 89.64% slowerfilter (empty array):
```
## Processing stage
### flatten (100x10x10):
**Before**
```js
  Rambda:
    433 ops/s, ±1.21%       | slowest, 99.59% slower
  Ramda:
    522 ops/s, ±2.35%       | 99.5% slower
  Lodash:
    104 409 ops/s, ±2.09%   | fastest
```
**After**
```js
  Rambda:
    11 127 ops/s, ±2.44%    | 90.21% slower
  Ramda:
    510 ops/s, ±2.64%       | slowest, 99.55% slower
  Lodash:
    113 605 ops/s, ±1.70%   | fastest
```