@selfrefactor 2020-06-04
Released with `5.6.1`